Adding compile_xeddis.P and making a condition for the test suite.
git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@7071 756869e0-9e47-4bc5-b0a8-847305dc5e1c

+++ xsbtests/prolog_tests/test.sh
@@ -7,8 +7,19 @@ echo "-------------------------------------------------------"
 XEMU=$1
 options=$2
 
-rm -f xeddis.dylib xeddis.so
-$XEMU -e "[compile_xeddis]."
+u=`uname`;
+echo "uname for this system is $u";
+
+if test $u  != ""; then
+    echo "removing xeddis object files"
+    rm -f xeddis.dylib xeddis.so
+    $XEMU -e "[compile_xeddis]."
+    echo "-------------------------------------------------------"
+    else
+    echo "not removing object files"
+    echo "-------------------------------------------------------"
+fi
+
 
 #------------------------------------
 # tests involving standard predicates
